requirements-basic.txt
# Core dependencies
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0

# Document processing
python-docx==1.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0